move mesh generation to Stitch class; remove Yarn class
