--- KnittingChartPreview/Assets/Scripts/Yarn.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1c76080af5de4415aaf142fcb78795aa
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
